% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATLdownload.R
\name{ATLdownload}
\alias{ATLdownload}
\title{Downloads ATLAS data products with a curl request}
\usage{
ATLdownload(
  short_name,
  bounding_box,
  bbox,
  shapefile,
  time,
  token,
  version,
  coverage,
  projection,
  tempDir,
  dwlDir,
  email = "yes"
)
}
\arguments{
\item{short_name}{Short name ID of the dataset. For example ATL03 or ATL08}

\item{bounding_box}{Bounding box used as a search filter. lower_left_long, lower_left_lat, upper_right_long, upper_righ_lat}

\item{bbox}{Bounding box used for spatial subsetting. Must match the coordinates entered for bounding_box. If left NULL, automatically copies bounding_box values.}

\item{shapefile}{An ESRI Shapefile used for spatial subsetting. Output files will be cropped to the shapefile extent.}

\item{time}{Temporal range filter}

\item{token}{Token required to access data using Eathdata login credentials}

\item{version}{Dataset version number. Default is 002}

\item{coverage}{Subsets dataset parameters (based on HDF5 dataset structure)}

\item{projection}{Projection of the output files.}

\item{tempDir}{Full path to temporary folder where download request and eventually clipping shapefile will be saved}

\item{dwlDir}{Full path to folder where data will be downloaded}
}
\description{
curl needs to be installed on your machine and a token available to download data
}
