% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ATL_naming}
\alias{check_ATL_naming}
\title{Check if file follows conventional ATLAS file naming}
\usage{
check_ATL_naming(file, product)
}
\description{
Check if file follows conventional ATLAS file naming
}
