% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_ATL03.R
\name{classify_ATL03}
\alias{classify_ATL03}
\title{Classify ATL03 photons to ATL08 classes}
\usage{
classify_ATL03(
  atl03_h5,
  atl08_h5,
  lat_range,
  beam = c("gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"),
  beam_strength = c("weak", "strong"),
  odir
)
}
\arguments{
\item{atl03_h5}{Path to ATL03 HDF file}

\item{atl08_h5}{Path to ATL08 file}

\item{lat_range}{Optional. Vector indicating the range of latitude to process}

\item{beam}{Character vector indicating beams to process}

\item{beam_strength}{Character vector indicating the strength of beams to process}

\item{odir}{Directory where ouput files will be written}
}
\description{
Classify ATL03 photons to ATL08 classes
}
