% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_atl08_shp.R
\name{combine_atl08_shp}
\alias{combine_atl08_shp}
\title{Combine ATL08 shp files}
\usage{
combine_atl08_shp(directory, out_dir = directory, out_type = "gpkg")
}
\arguments{
\item{out_type}{}
}
\description{
Combine ATL08 shp files
}
