% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_atl_pairs}
\alias{find_atl_pairs}
\title{Find pairs of ATL03 and ATL08 files}
\usage{
find_atl_pairs(atl03_flist, atl08_flist)
}
\arguments{
\item{atl03_flist}{Vector of character listing the file path to ATL03 HDF files}

\item{atl08_flist}{Vector of character listing the file path to ATL08 HDF files}
}
\description{
Find pairs of ATL03 and ATL08 files
}
