% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_along_distance}
\alias{get_along_distance}
\title{Convert x/y to along and across distance}
\usage{
get_along_distance(x_in, y_in, R_mat, xRotPt, yRotPt)
}
\description{
Convert x/y to along and across distance
}
