% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getATL03ph.R
\name{getATL03ph}
\alias{getATL03ph}
\title{Extract individual photons from an ATL03 file}
\usage{
getATL03ph(
  file,
  beam = c("gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"),
  beam_strength = c("weak", "strong"),
  lat_range,
  seg_filter,
  id_sep = "_"
)
}
\arguments{
\item{seg_filter}{Vector of integers. Filter by 20 m segments of the ATL03 product}

\item{id_sep}{}
}
\description{
Returns a data.frame with lat, lon and height of each photon and their associated segement and unique ID
}
