% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getATL03Radiometry.R
\name{getATL03Radiometry}
\alias{getATL03Radiometry}
\title{Radiometry of ATL03}
\usage{
getATL03Radiometry(ATL03_df, signal)
}
\arguments{
\item{ground}{}
}
\description{
Compute the radiometry, defined as the number of signal photons per outgoing ATLAS laser shot
}
