% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getATL08ph.R
\name{getATL08ph}
\alias{getATL08ph}
\title{Extract individual classified photons from an ATL08 file}
\usage{
getATL08ph(
  file,
  beam = c("gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"),
  beam_strength = c("weak", "strong"),
  lat_range,
  seg_filter,
  id_sep = "_",
  flags = c("snow", "night", "msw")
)
}
\arguments{
\item{seg_filter}{Vector of integers. Filter by 20 m segments of the ATL03 product}

\item{flags}{Character. ATL08 segement-level flag to return. All photons within the segments will be assigned the same flag. Impemented flags are: "snow", "night", "msw".}
}
\description{
Returns a data.frame with lat, lon and height of each photon and their associated segement and unique ID
}
