% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getATL08Segments.R
\name{getATL08Segments}
\alias{getATL08Segments}
\title{Get the segment of each ATL08 point}
\usage{
getATL08Segments(
  file,
  beam = c("gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"),
  beam_strength = c("weak", "strong"),
  crs_proj = 26917,
  odir
)
}
\arguments{
\item{file}{Path to ATL03 or ATL08 HDF5 file (.h5 extension)}

\item{beam}{Character. List of beams for which segments to be extracted}

\item{rgt}{Logical. Should RGT number be added as a field to output feature?}

\item{cycle}{Logical. Should cycle number be added as a field to output feature?}

\item{segment}{Logical. Should segment number be added as a field to output feature?}

\item{date}{Logical. Should date be added as a field to output feature?}

\item{time}{Logical. Should time be added as a field to output feature?}

\item{...}{Parameters passed to writeOGR (e.g. overwrite_layer = TRUE)}
}
\description{
Get the segment of each ATL08 point
}
