% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroundTrack.R
\name{getGroundTrack}
\alias{getGroundTrack}
\title{Get the ground track of ATLAS beams based on ATL03 ot ATL08 data product}
\usage{
getGroundTrack(
  file,
  product,
  beam = c("gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"),
  beam_strength = c("weak", "strong"),
  points = FALSE,
  lines = TRUE,
  df = TRUE,
  write = TRUE,
  outDir,
  filename_points,
  filename_lines,
  ...
)
}
\arguments{
\item{file}{Path to ATL03 or ATL08 HDF5 file (.h5 extension)}

\item{product}{Either "ATL03" or "ATL08"}

\item{beam}{Character. List of beams for which groudn track needs to be extracted}

\item{points}{Logical. Should each reference point be returned in a SpatialPointsDataFrame?}

\item{lines}{Logical. Should lines features between all reference points be returned in a SpatialLinesDataFrame?}

\item{...}{Parameters passed to writeOGR (e.g. overwrite_layer = TRUE)}

\item{rgt}{Logical. Should RGT number be added as a field to output feature?}

\item{cycle}{Logical. Should cycle number be added as a field to output feature?}

\item{segment}{Logical. Should segment number be added as a field to output feature?}

\item{date}{Logical. Should date be added as a field to output feature?}

\item{time}{Logical. Should time be added as a field to output feature?}
}
\description{
Get the ground track of ATLAS beams based on ATL03 ot ATL08 data product
}
