% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighestCanopyPhoton.R
\name{HighestPhotonWindow}
\alias{HighestPhotonWindow}
\title{Detect the highest canopy photon within moving window}
\usage{
HighestPhotonWindow(
  df,
  window = 0.7,
  h_norm = "h_norm",
  beam = c("gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"),
  gap_dist = 10,
  min_ph_gap = 20,
  keep_cols,
  dist_along = "dist_along",
  ph_id = "ph_id",
  beam_n = "beam",
  ATL08_class = "ph_class"
)
}
\arguments{
\item{window}{}
}
\description{
Detect the highest canopy photon within moving window
}
