% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{leaflet_base_map}
\alias{leaflet_base_map}
\title{a leaflet base map}
\usage{
leaflet_base_map(
  sf_obj = NULL,
  leaflet_provider_base,
  leaflet_provider_secondary,
  overlay_group = "IceSat2 1-degree",
  overlay_polylines = "Polylines",
  width = NULL,
  height = NULL,
  popup = NULL,
  collapse_layer_control = TRUE
)
}
\arguments{
\item{sf_obj}{a simple features object which must include the columns 'date' and 'size_MB' if the popup parameter is set to TRUE}

\item{leaflet_provider_base}{a leaflet provider object}

\item{leaflet_provider_secondary}{a leaflet provider object}

\item{overlay_group}{a character string specifying the overlay group}

\item{overlay_polylines}{a character string specifying the Polylines group}

\item{width}{a numeric value specifying the width of the leaflet map}

\item{height}{a numeric value specifying the height of the leaflet map}

\item{popup}{a boolean specifying if the popup should be shown or not}

\item{collapse_layer_control}{a boolean specifying if the layer-control should be collapsed or not}
}
\value{
a leaflet object (map)
}
\description{
a leaflet base map
}
\keyword{internal}
