% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transect.R
\name{make_transect}
\alias{make_transect}
\title{Generate LINESTRING from ATL08 center points}
\usage{
make_transect(atl08_sf, out_dir, tr_length = 100)
}
\arguments{
\item{out_dir}{Directory where output will be written (as gpkg file)}

\item{tr_length}{Length of transect (DO NOT CHANGE)}

\item{atl08_df}{Data.frame or sf object containing the center points coordinates of ATL08 100-m segments. Can be generated from \link[icesat2R]{read_ATL08}}
}
\description{
Generate LINESTRING from ATL08 center points
}
