% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeH.R
\name{normalizeH}
\alias{normalizeH}
\title{Normalize photons height}
\usage{
normalizeH(
  df,
  h = "h",
  beam = c("gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"),
  gap_dist = 10,
  min_ph_gap = 20,
  method = "loess",
  returnGaps = TRUE,
  keep_cols,
  dist_along = "cum_dist_along",
  segment_id = "seg_id",
  segment_idx = "seg_idx",
  ph_id = "ph_id",
  beam_n = "beam",
  ATL08_class = "ph_class"
)
}
\arguments{
\item{df}{Data frame}

\item{h}{x Name of df column indicating photon height}

\item{x}{Name of df column indicating photon x coordinate}

\item{y}{Name of df column indicating photon y coordinate}

\item{dh}{min and max height thresholds used to flag unsual photons height}

\item{dx}{min and max horizontal distance}
}
\description{
Use local polynomial regression fitting with degree 0 (moving average) to interpolate ground surface and normalize photons using the interpolated surface
}
\details{
Segments with continuous ground points are first identified and interpolation is performaed at these segment level. THe continuity between ground point is determined by the gap_dist threshold: if two groudn points are more than gap_dist along-distance apart, a gap is detected and ground surface will be interpolated for each gap. Additionally, segments with less than min_ph_gap photons in them are discarded. Therefore, not all photons in the original df are guaranteed to be normalized.
}
