% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ATL03.R
\name{read_ATL03}
\alias{read_ATL03}
\title{Read ATL03 h5 file as a data frame}
\usage{
read_ATL03(
  file,
  beam = c("gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"),
  beam_strength = c("weak", "strong"),
  lat_range,
  odir_atl03
)
}
\arguments{
\item{file}{Path to h5 file}

\item{beam}{Character vector indicating beams to process}

\item{beam_strength}{Character vector indicating the strength of beams to process}

\item{lat_range}{Numeric vector. Lower and upper latitude to return}

\item{odir_atl03}{Character. Output directory of ATL03 product}
}
\description{
Read ATL03 h5 file as a data frame
}
