% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_utils.R
\name{switch_full}
\alias{switch_full}
\title{Convert a full month name to a numeric value}
\usage{
switch_full(month)
}
\arguments{
\item{month}{a character string specifying one of the twelve months, i.e. ("January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December")}
}
\value{
a character string which will correspond to the numeric value of the input month, i.e. ("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12")
}
\description{
Convert a full month name to a numeric value
}
\keyword{internal}
