% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_fill.R
\name{vegetationFill}
\alias{vegetationFill}
\title{Calculates vegetation fill index}
\usage{
vegetationFill(
  df,
  th,
  dist_along = "dist_along",
  h_norm = "h_norm",
  top_ph = "top_ph",
  ph_class = "ph_class"
)
}
\arguments{
\item{ph_class}{}
}
\description{
Calculates vegetation fill index
}
