% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroundTrack.R
\name{getGroundTrack}
\alias{getGroundTrack}
\title{Get the ground track of ATLAS beams based on ATL03 data product}
\usage{
getGroundTrack(ATL03, beam = c("gt1l", "gt1r", "gt2l", "gt2r", "gt3l",
  "gt3r"), rgt = TRUE, cycle = TRUE, segment = TRUE, date = TRUE,
  time = TRUE, points = FALSE, lines = TRUE, write = TRUE, outDir,
  filename_points, filename_lines, ...)
}
\arguments{
\item{ATL03}{Path to ATL03 HDF5 file (.h5 extension)}

\item{beam}{Character. List of beams for which groudn track needs to be extracted}

\item{rgt}{Logical. Should RGT number be added as a field to output feature?}

\item{cycle}{Logical. Should cycle number be added as a field to output feature?}

\item{segment}{Logical. Should segment number be added as a field to output feature?}

\item{date}{Logical. Should date be added as a field to output feature?}

\item{time}{Logical. Should time be added as a field to output feature?}

\item{points}{Logical. Should each reference point be returned in a SpatialPointsDataFrame?}

\item{lines}{Logical. Should lines features between all reference points be returned in a SpatialLinesDataFrame?}

\item{...}{Parameters passed to writeOGR (e.g. overwrite_layer = TRUE)}
}
\description{
Get the ground track of ATLAS beams based on ATL03 data product
}
