% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mission_orbits.R
\name{latest_orbits}
\alias{latest_orbits}
\title{Extracting the download url's from the Technical Specs website}
\usage{
latest_orbits(
  technical_specs_url = "https://icesat-2.gsfc.nasa.gov/science/specs",
  verbose = FALSE
)
}
\arguments{
\item{technical_specs_url}{a character string specifying the technical specs website}

\item{verbose}{a boolean. If TRUE then information will be printed out in the console}
}
\value{
a 'data.table' object
}
\description{
Extracting the download url's from the Technical Specs website
}
\examples{

\dontrun{

require(IceSat2R)

orbs = latest_orbits(verbose = TRUE)
orbs

}
}
\references{
https://icesat-2.gsfc.nasa.gov/science/specs
}
